updates nft profile context to export `unnamedUser`; adds `fetchUserLikes` and `likeDislike` methods to set sessionUser current likes
