--- src/pages/NFTs/Profile/unnamedUser.ts
@@ -1,13 +1,0 @@
-export const unnamedUser = {
-  user_id: null,
-  pubkey: '',
-  nickname: 'Unnamed',
-  email: '',
-  bio: '',
-  twitter_link: '',
-  instagram_link: '',
-  facebook_link: '',
-  youtube_link: '',
-  profile_pic_link: '',
-  is_verified: false
-}
